rename serialNumberTransCountsActions
