+++ src/Service.Host/client/src/containers/serialNumberTransactions/CreateSerialNumberTransaction.js
@@ -1,7 +1,7 @@
 ﻿import { connect } from 'react-redux';
 import { fetchErrorSelectors } from '@linn-it/linn-form-components-library';
 import serialNumberTransactionActions from '../../actions/serialNumberTransactionActions';
-import serialNumberTransactionUtilityActions from '../../actions/serialNumberTransactionUtilityActions';
+import serialNumberTransCountsActions from '../../actions/serialNumberTransCountsActions';
 import SerialNumberTransaction from '../../components/serialNumberTransactions/SerialNumberTransaction';
 import serialNumberTransactionSelectors from '../../selectors/serialNumberTransactionSelectors';
 import serialNumberTransactionCountsSelectors from '../../selectors/serialNumberTransactionCountsSelectors';
@@ -17,7 +17,7 @@ const mapStateToProps = state => ({
 });
 
 const initialise = () => dispatch => {
-    dispatch(serialNumberTransactionUtilityActions.fetch());
+    dispatch(serialNumberTransCountsActions.fetch());
 };
 
 const mapDispatchToProps = {
